ath11k: Compilation warining fixes for ath11k Platform files
qmi.h  -  Modified the data type of addr to u32 * in m3_dump_data structure
hw.c   -  Updated proper struct hal_rx_mpdu_info_qcn9074 in place of hal_rx_mpdu_info_qcn9074
          Modified function param to struct hal_rx_mpdu_info* instead of u8*
core.c -  Modified the function param of enable_smart_antenna from int to bool
core.h -  Modified the data type of adfs_chain_mask to unsigned long in ath11k_pdev_cap structure

+++ backports-6.1-9a0dddfb3/drivers/net/wireless/ath/ath11k/core.h
@@ -1149,7 +1149,7 @@ struct ath11k_pdev_cap {
 	u32 rx_chain_mask;
 	u32 tx_chain_mask_shift;
 	u32 rx_chain_mask_shift;
-	u32 adfs_chain_mask;
+	unsigned long adfs_chain_mask;
 	struct ath11k_band_cap band[NUM_NL80211_BANDS];
 	bool nss_ratio_enabled;
 	u8 nss_ratio_info;
